Rename and added more info
